Changed the heading buttons in lite mode to say "(Refresh) (Search) (Return to Full Mode)"
One the main reasons I have to leave lite-mode is to search, so it just makes sense to surface a
Search link at the top of lite-mode.

Also, I changed (Return to Full Mode) to use display_mode=0 and not 2.  Display mode 2 is
basically useless and it doesn't make sense to link to it.

* forum.php: